genai/pipeline/functions/collector/requirements.txt
functions-framework==3.*
duckdb==1.1.0
google-cloud-secret-manager
pandas